+++ modules/juce_audio_processors/processors/juce_AudioProcessor.h
@@ -764,7 +764,7 @@ public:
     }
 
     template <int numLayouts>
-    static bool containsLayout (const BusesLayout& layouts, const short (&channelLayoutList) [numLayouts][2])
+    static bool containsLayout (const BusesLayout& layouts, const short (&channelLayoutList) [(size_t) numLayouts][2])
     {
         return containsLayout (layouts, layoutListToArray (channelLayoutList));
     }
@@ -782,7 +782,7 @@ public:
     */
     template <int numLayouts>
     BusesLayout getNextBestLayoutInLayoutList (const BusesLayout& layouts,
-                                               const short (&channelLayoutList) [numLayouts][2])
+                                               const short (&channelLayoutList) [(size_t) numLayouts][2])
     {
         return getNextBestLayoutInList (layouts, layoutListToArray (channelLayoutList));
     }
@@ -1422,7 +1422,7 @@ private:
     };
 
     template <int numLayouts>
-    static Array<InOutChannelPair> layoutListToArray (const short (&configuration) [numLayouts][2])
+    static Array<InOutChannelPair> layoutListToArray (const short (&configuration) [(size_t) numLayouts][2])
     {
         Array<InOutChannelPair> layouts;
